Add basic Audio support for Pepper
Pepper uses the TLV320AIC3106 Audio Codec via McASP0.  This is clocked
by a 12MHz oscillator and uses the AXR0 and AXR1 data lanes. The headset
jack is connected to the HPLOUT and HPROUT outputs.

Note: Bluetooth PCM is also connected to the audio codec but is not
enabled in this commit.